requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.9.24
charset-normalizer==2.1.1
cloudscraper==1.2.64
config==0.5.1
idna==3.4
Js2Py==0.71
pyjsparser==2.7.1
pyparsing==3.0.9
pyTelegramBotAPI==4.7.0
python-aternos==2.1.1
python-dotenv==0.21.0
python-telegram-bot==13.14
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
regex==2022.9.13
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
telebot==0.0.4
tornado==6.1
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
utils==1.0.1
websockets==10.3
